--- data/.gitignore
@@ -1,1 +1,0 @@
-fiat.py
