+++ new_albums/classes/artistClass.py
