+++ new_albums/config.py
@@ -8,8 +8,9 @@ from decouple import config as dconfig
 PLAYLIST_ID = dconfig("NEW_ALBUMS_PLAYLIST_ID")
 SPOTIFY_CLIENT_ID = dconfig("SPOTIFY_CLIENT_ID")
 SPOTIFY_CLIENT_SECRET = dconfig("SPOTIFY_CLIENT_SECRET")
-SPOTIFY_REDIRECT_URI = dconfig("SPOTIFY_REDIRECT_URI")
+SPOTIFY_REDIRECT_URI = dconfig("SPOTIFY_REDIRECT_URI", default="http://localhost:8080")
 SPOTIFY_USER = dconfig("SPOTIFY_USER")
+FIAT_FILE = dconfig("FIAT_FILE", default="_default_fiat")
 
 
 # The rest are also config settings, but their literal values are commited to the codebase since they aren't secret.
